# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GamingOS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, top_rated, trending, price_asc, price_desc] }
        - name: subscription_tier
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a game (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Users ───────────────────────────────────────────────────────────────────
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions (admin)
      parameters:
        - name: plan
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Subscribe a user to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription (change plan/status)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ── Reviews ─────────────────────────────────────────────────────────────────
  /games/{id}/reviews:
    get:
      operationId: listGameReviews
      tags: [reviews]
      summary: List reviews for a game
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a game
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin/author)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Wishlist ─────────────────────────────────────────────────────────────────
...
  /achievements:
    get:
      operationId: listAchievements
      tags: [achievements]
      summary: List all achievements
      parameters:
        - name: gameId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"
    post:
      operationId: createAchievement
      tags: [achievements]
      summary: Create an achievement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AchievementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
  /users/{id}/achievements:
    get:
      operationId: getUserAchievements
      tags: [achievements]
      summary: Get achievements unlocked by user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"

  # ── Store summary (wow) ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Game:
      type: object
      required: [id, title, description, category, price, rating, downloads, coverImage, status, subscriptionTier, isFeatured, isTrending, releaseDate, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        rating: { type: number }
        reviewCount: { type: integer }
        downloads: { type: integer }
        coverImage: { type: string }
        trailerUrl: { type: ["string", "null"] }
        screenshots:
          type: array
          items: { type: string }
        developer: { type: ["string", "null"] }
        publisher: { type: ["string", "null"] }
        status: { type: string, enum: [active, coming_soon, removed] }
        subscriptionTier: { type: string, enum: [free, basic, premium, vip] }
        isFeatured: { type: boolean }
        isTrending: { type: boolean }
        releaseDate: { type: string, format: date }
        tags:
          type: array
          items: { type: string }
        requirements: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    GameInput:
      type: object
      required: [title, description, category, price, coverImage, subscriptionTier]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        coverImage: { type: string }
        trailerUrl: { type: string }
        developer: { type: string }
        publisher: { type: string }
        subscriptionTier: { type: string, enum: [free, basic, premium, vip] }
        isFeatured: { type: boolean }
        isTrending: { type: boolean }
        releaseDate: { type: string, format: date }
        tags:
          type: array
          items: { type: string }
        requirements: { type: string }
        status: { type: string, enum: [active, coming_soon, removed] }
    GameUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        coverImage: { type: string }
        trailerUrl: { type: string }
        developer: { type: string }
        publisher: { type: string }
        subscriptionTier: { type: string, enum: [free, basic, premium, vip] }
        isFeatured: { type: boolean }
        isTrending: { type: boolean }
        status: { type: string, enum: [active, coming_soon, removed] }
        tags:
          type: array
          items: { type: string }
...
    Subscription:
      type: object
      required: [id, userId, plan, status, billingCycle, startDate, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        username: { type: ["string", "null"] }
        plan: { type: string, enum: [free, basic, premium, vip] }
        status: { type: string, enum: [active, cancelled, expired, trial] }
        billingCycle: { type: string, enum: [monthly, quarterly, yearly] }
        price: { type: number }
        startDate: { type: string, format: date }
        endDate: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    SubscriptionInput:
      type: object
      required: [userId, plan, billingCycle]
      properties:
        userId: { type: integer }
        plan: { type: string, enum: [free, basic, premium, vip] }
        billingCycle: { type: string, enum: [monthly, quarterly, yearly] }
    SubscriptionUpdate:
      type: object
      properties:
        plan: { type: string, enum: [free, basic, premium, vip] }
        status: { type: string, enum: [active, cancelled, expired, trial] }
        billingCycle: { type: string, enum: [monthly, quarterly, yearly] }
        endDate: { type: string, format: date }
    Review:
      type: object
      required: [id, gameId, userId, username, rating, content, createdAt]
      properties:
        id: { type: integer }
        gameId: { type: integer }
        userId: { type: integer }
        username: { type: string }
        avatarUrl: { type: ["string", "null"] }
        rating: { type: integer }
        content: { type: string }
        helpful: { type: integer }
        createdAt: { type: string, format: date-time }
    ReviewInput:
      type: object
      required: [userId, rating, content]
      properties:
        userId: { type: integer }
        rating: { type: integer, minimum: 1, maximum: 5 }
        content: { type: string, minLength: 10 }
...
    Achievement:
      type: object
      required: [id, name, description, xpReward, rarity]
      properties:
        id: { type: integer }
        gameId: { type: ["integer", "null"] }
        gameName: { type: ["string", "null"] }
        name: { type: string }
        description: { type: string }
        iconUrl: { type: ["string", "null"] }
        xpReward: { type: integer }
        rarity: { type: string, enum: [common, rare, epic, legendary] }
        unlockedAt: { type: ["string", "null"] }
    AchievementInput:
      type: object
      required: [name, description, xpReward, rarity]
      properties:
        gameId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string }
        iconUrl: { type: string }
        xpReward: { type: integer }
        rarity: { type: string, enum: [common, rare, epic, legendary] }
...