# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GamingOS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from './gameStatus';
import type { GameSubscriptionTier } from './gameSubscriptionTier';

export interface Game {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  rating: number;
  reviewCount?: number;
  downloads: number;
  coverImage: string;
  /** @nullable */
  trailerUrl?: string | null;
  screenshots?: string[];
  /** @nullable */
  developer?: string | null;
  /** @nullable */
  publisher?: string | null;
  status: GameStatus;
  subscriptionTier: GameSubscriptionTier;
  isFeatured: boolean;
  isTrending: boolean;
  releaseDate: Date;
  tags?: string[];
  /** @nullable */
  requirements?: string | null;
  createdAt: Date;
}